refactor url
