+++ packages/expo-cli/src/commands/url/urlAsync.ts
@@ -4,8 +4,8 @@ import { ProjectSettings, UrlUtils } from 'xdl';
 
 import CommandError from '../../CommandError';
 import Log from '../../log';
-import printRunInstructionsAsync from '../../printRunInstructionsAsync';
 import urlOpts, { URLOptions } from '../../urlOpts';
+import printRunInstructionsAsync from './printRunInstructionsAsync';
 
 type ProjectUrlOptions = Command & {
   web?: boolean;
